+++ source/extensions/isaacsim.core.nodes/python/tests/test_core_nodes.py
@@ -16,7 +16,7 @@
 
 import carb
 import omni.kit.test
-from isaacsim.core.nodes.bindings import _isaacsim_core_nodes
+from isaacsim.core.simulation_manager import SimulationManager
 from isaacsim.core.utils.stage import open_stage_async
 from isaacsim.storage.native import get_assets_root_path_async
 
@@ -25,7 +25,6 @@ class TestCoreNodes(omni.kit.test.AsyncTestCase):
     async def setUp(self):
         """Set up  test environment, to be torn down when done"""
         self._timeline = omni.timeline.get_timeline_interface()
-        self._core_nodes = _isaacsim_core_nodes.acquire_interface()
         # add franka robot for test
         assets_root_path = await get_assets_root_path_async()
         if assets_root_path is None:
@@ -43,34 +42,34 @@ class TestCoreNodes(omni.kit.test.AsyncTestCase):
     # ----------------------------------------------------------------------
     async def test_simulation_time(self):
         await omni.kit.app.get_app().next_update_async()
-        a = self._core_nodes.get_sim_time()
-        b = self._core_nodes.get_sim_time_monotonic()
-        c = self._core_nodes.get_system_time()
+        a = SimulationManager.get_simulation_time()
+        b = SimulationManager._simulation_manager_interface.get_simulation_time_monotonic()
+        c = SimulationManager._simulation_manager_interface.get_system_time()
         await omni.kit.app.get_app().next_update_async()
-        a = self._core_nodes.get_sim_time_at_time((0, 1))
-        b = self._core_nodes.get_sim_time_monotonic_at_time((0, 1))
-        c = self._core_nodes.get_system_time_at_time((0, 1))
+        a = SimulationManager._simulation_manager_interface.get_simulation_time_at_time((0, 1))
+        b = SimulationManager._simulation_manager_interface.get_simulation_time_monotonic_at_time((0, 1))
+        c = SimulationManager._simulation_manager_interface.get_system_time_at_time((0, 1))
         self._timeline.play()
         await omni.kit.app.get_app().next_update_async()
-        a = self._core_nodes.get_sim_time()
-        b = self._core_nodes.get_sim_time_monotonic()
-        c = self._core_nodes.get_system_time()
-        a = self._core_nodes.get_sim_time_at_time((0, 0))
-        b = self._core_nodes.get_sim_time_monotonic_at_time((0, 0))
-        c = self._core_nodes.get_system_time_at_time((0, 0))
+        a = SimulationManager.get_simulation_time()
+        b = SimulationManager._simulation_manager_interface.get_simulation_time_monotonic()
+        c = SimulationManager._simulation_manager_interface.get_system_time()
+        a = SimulationManager._simulation_manager_interface.get_simulation_time_at_time((0, 0))
+        b = SimulationManager._simulation_manager_interface.get_simulation_time_monotonic_at_time((0, 0))
+        c = SimulationManager._simulation_manager_interface.get_system_time_at_time((0, 0))
 
     # ----------------------------------------------------------------------
     async def test_physics_num_steps(self):
-        steps = self._core_nodes.get_physics_num_steps()
+        steps = SimulationManager.get_num_physics_steps()
         self.assertEqual(steps, 0)
         self._timeline.play()
         await omni.kit.app.get_app().next_update_async()
-        steps = self._core_nodes.get_physics_num_steps()
+        steps = SimulationManager.get_num_physics_steps()
         self.assertEqual(steps, 3)
         await omni.kit.app.get_app().next_update_async()
-        steps = self._core_nodes.get_physics_num_steps()
+        steps = SimulationManager.get_num_physics_steps()
         self.assertEqual(steps, 4)
         self._timeline.stop()
         await omni.kit.app.get_app().next_update_async()
-        steps = self._core_nodes.get_physics_num_steps()
+        steps = SimulationManager.get_num_physics_steps()
         self.assertEqual(steps, 0)